Move prefetch logic to an util; Leverage existing nextMediaItem
